Adecuación de Grabado de Tarifas.

+++ src/main/java/pe/com/isesystem/siscopetarifario/model/TarifarioGeneralId.java
@@ -26,7 +26,7 @@ public class TarifarioGeneralId implements Serializable {
     private TipoServicio idTipoServicio;
 
     @Column(name = "id_dia", nullable = false)
-    private Integer idDia;
+    private Long idDia;
 
     @Override
     public boolean equals(Object o) {
